requirements.txt
flask
yt-dlp
moviepy
pydub
gunicorn
python-dotenv
